+++ training-service/src/main/java/fr/hoenheimsports/trainingservice/assemblers/TeamAssemblerImpl.java
@@ -7,6 +7,7 @@ import fr.hoenheimsports.trainingservice.mappers.TeamMapper;
 import fr.hoenheimsports.trainingservice.models.Category;
 import fr.hoenheimsports.trainingservice.models.Gender;
 import fr.hoenheimsports.trainingservice.models.Team;
+import fr.hoenheimsports.trainingservice.services.UserSecurityService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PagedResourcesAssembler;
@@ -25,10 +26,12 @@ public class TeamAssemblerImpl implements TeamAssembler {
 
     private final TeamMapper teamMapper;
     private final PagedResourcesAssembler<Team> pagedResourcesAssembler;
+    private final UserSecurityService userSecurityService;
 
-    public TeamAssemblerImpl(TeamMapper teamMapper, PagedResourcesAssembler<Team> pagedResourcesAssembler) {
+    public TeamAssemblerImpl(TeamMapper teamMapper, PagedResourcesAssembler<Team> pagedResourcesAssembler, UserSecurityService userSecurityService) {
         this.teamMapper = teamMapper;
         this.pagedResourcesAssembler = pagedResourcesAssembler;
+        this.userSecurityService = userSecurityService;
     }
 
 
@@ -57,17 +60,25 @@ public class TeamAssemblerImpl implements TeamAssembler {
     }
 
     private void addLinks(CollectionModel<TeamDTO> resources) {
-        resources.add(linkTo(methodOn(TeamControllerImpl.class).getAllTeams(Pageable.unpaged())).withSelfRel()
-                .andAffordance(afford(methodOn(TeamControllerImpl.class).createTeam(new TeamDTORequest(1L, Gender.N, Category.SENIOR,1, Set.of(), Set.of())))) // skip default name
-                .andAffordance(afford(methodOn(TeamControllerImpl.class).createTeam(new TeamDTORequest(1L, Gender.N, Category.SENIOR,1,Set.of(), Set.of()))))
-        );
+        resources.add(linkTo(methodOn(TeamControllerImpl.class).getAllTeams(Pageable.unpaged())).withSelfRel());
+        if(this.userSecurityService.hasRole("ADMIN")) {
+            resources.add(linkTo(methodOn(TeamControllerImpl.class).getAllTeams(Pageable.unpaged())).withSelfRel()
+                    .andAffordance(afford(methodOn(TeamControllerImpl.class).createTeam(new TeamDTORequest(1L, Gender.N, Category.SENIOR,1, Set.of(), Set.of())))) // skip default name
+                    .andAffordance(afford(methodOn(TeamControllerImpl.class).createTeam(new TeamDTORequest(1L, Gender.N, Category.SENIOR,1,Set.of(), Set.of()))))
+            );
+        }
+
     }
 
     private void addLinks(TeamDTO resource) {
-        resource.add(linkTo(methodOn(TeamControllerImpl.class).getTeamById(resource.getId())).withSelfRel()
-                .andAffordance(afford(methodOn(TeamControllerImpl.class).updateTeam(resource.getId(), null))) //skip default
-                .andAffordance(afford(methodOn(TeamControllerImpl.class).updateTeam(resource.getId(), null)))
-                .andAffordance(afford(methodOn(TeamControllerImpl.class).deleteTeam(resource.getId()))));
+        resource.add(linkTo(methodOn(TeamControllerImpl.class).getTeamById(resource.getId())).withSelfRel());
         resource.add(linkTo(methodOn(TeamControllerImpl.class).getAllTeams(Pageable.unpaged())).withRel("teams").expand());
+        if(this.userSecurityService.hasRole("ADMIN")) {
+            resource.add(linkTo(methodOn(TeamControllerImpl.class).getTeamById(resource.getId())).withSelfRel()
+                    .andAffordance(afford(methodOn(TeamControllerImpl.class).updateTeam(resource.getId(), null))) //skip default
+                    .andAffordance(afford(methodOn(TeamControllerImpl.class).updateTeam(resource.getId(), null)))
+                    .andAffordance(afford(methodOn(TeamControllerImpl.class).deleteTeam(resource.getId()))));
+        }
+
     }
 }
